split the script into processing and keyword extracion, and also created the package
